migrate models to separate files
